move views into its own folder for easier updating
