+++ PathKit/src/components/cards/EntityCard.tsx
@@ -1,6 +1,6 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import classNames from "classnames";
-import { EntityObjectProps } from "../../CardView";
+import { EntityObjectProps } from "../views/CardView";
 import styles from "./EntityCard.module.scss";
 
 interface EntityCardProps {
