+++ PathKit/src/components/menus/PlannerMenu.module.scss
@@ -32,6 +32,7 @@
     font-weight: bold;
     text-decoration: none;
     color: #333;
+    cursor: "pointer";
 
     &:hover {
       color: #06d85f;
